+++ poc/classes/AuxSTVarSpecList.py
@@ -1,6 +1,7 @@
 from poc.classes.AuxEvaluation import EvaluateParams
 from poc.classes.AuxST import AuxSTOutline
 from poc.classes.AuxSTConstants import AuxSTConstants
+from poc.classes.AuxSTTypeInterface import AuxSTTypeInterface
 
 
 class AuxSTVarSpecList(AuxSTOutline):
@@ -13,7 +14,11 @@ class AuxSTVarSpecList(AuxSTOutline):
 
     def evaluate(self, sem):
         for child in self.children:
-            EvaluateParams.evaluate_recursion(sem, child)
+            if isinstance(child, AuxSTTypeInterface):
+                # call evaluation differently for all children that have predefined types
+                EvaluateParams.evaluate_recursion(sem, child, expected_type=child.get_declared_type())
+            else:
+                EvaluateParams.evaluate_recursion(sem, child)
 
     def get_long_id(self):
         if self._long_id is None: